refactor(extensions): the manifest records governed autonomy tiers, not every declaration
The manifest is a governance artifact (§5 feeds §7): it must list only
what an operator resolves — the operations an extension adds that run at
a 🟢/🟡 autonomy tier or request a scope. Passive policy an extension
supplies is not one of those: a jurisdiction pack is consulted by the
core, never invoked, and asks for no tier, so it contributes nothing to
the manifest. A jurisdiction-only unit (de, the crm-hello fixture) now
carries an empty autonomy_tiers list — there is genuinely nothing to
approve until a governed kind (agent tools, /x/ endpoints, ext tables)
lands.

This removes the jurisdiction/retention AST derivation entirely (it was
deriving passive policy into a governed surface it does not belong in),
collapsing the reader to: parse New(), read Name/Version, recognize and
skip Jurisdictions, and FAIL CLOSED on any unrecognized field — a future
governed field cannot ship as a silently empty manifest. The security-
descriptor shape (id, operation, scopes, tier, digest) stays as the
autonomy_tiers element type, ready for the first governed kind to emit
and digest its own requests.

Also resolves the Sonar S3776 cognitive-complexity findings on the
deleted parse functions.
