+++ docs/how-to/add-an-extension.md
@@ -115,11 +115,13 @@ have to regenerate the composition and run the gates:
 
 1. **`make composition`** — regenerates `build/composition/` from `extensions/`; your unit now appears
    in the generated `Extensions()`, and a `manifest.generated.json` lands next to your unit — the
-   statically derived record of what it declares. Commit it with the unit; the drift gate fails a
-   stale or hand-edited one. Derivation reads your `New()` (and the pack methods it references) from
-   the AST, so declarations must be literal values — a computed name, code, or period fails
-   generation with the file and line. (Every build/test lane depends on this target, so `make check`
-   runs it for you; run it directly when you want to inspect the output.)
+   statically derived record of the **autonomy tiers** it requests (the 🟢/🟡 operations and scopes an
+   operator must approve under §7; a jurisdiction-only unit requests none, so its list is empty).
+   Commit it with the unit; the drift gate fails a stale or hand-edited one. Derivation reads your
+   `New()` from the AST, so it must be a literal — a computed value, or a field the generator does not
+   recognize, fails generation with the file and line rather than silently dropping a request. (Every
+   build/test lane depends on this target, so `make check` runs it for you; run it directly when you
+   want to inspect the output.)
 2. **`make check`** — builds the composed workspace, runs the extension-tier fitness tests
    (import-boundary, marker placement, composition wiring), `make test-extensions` (your unit's own
    tests), and `make check-composition` (a clean regeneration must reproduce `composition.json`
